(maint) Add ffi-win32-extensions gem
This is a dependency that was added to later versions of the
win32-service gem. It defines a method with the same name as one in
the api_types.rb file in puppet, but as long as we require this gem
first prior to api_types it should not clash.

The method itself is not used in win32-service.